% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqq.R
\name{fastqq}
\alias{fastqq}
\title{fastqq}
\usage{
fastqq(data, p = "P", lambda = T, main = "Q-Q plot")
}
\arguments{
\item{data}{A GWAS summary with at least a column of P-value.}

\item{p}{A string for the header for the column of P-value in data.}

\item{lambda}{Set FALSE to exclude lambda from the plot.}

\item{main}{A string for the title of the plot.}
}
\description{
Creates Q-Q plots from GWAS summaries.
}

