% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata-data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{This is data to be included in my package}
\format{An object of class \code{data.frame} with 309579 rows and 3 columns.}
\usage{
testdata
}
\description{
This is data to be included in my package
}
\author{
Daniel Liang-Dar Hwang \email{danielldhwang@gmail.com}
}
\keyword{data}

