% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastman.R
\name{fastman}
\alias{fastman}
\title{fastman}
\usage{
fastman(data, chr = "CHR", ps = "BP", p = "P", main = "Manhattan plot",
  suggest_line = -log10(1e-05), gws_line = -log10(5e-08),
  color = c("grey", "black"), chr_build = "GRCh37", yscale = NA,
  xlab_all = F, turbo = F)
}
\arguments{
\item{data}{A GWAS summary with columns of chromosome number (X and Y chromosomes need to be replaced by 23 and 24), position (base pair) information, and P-value. ALL values have to be numeric.}

\item{chr}{A string for the header of the column of chromosome number in data.}

\item{ps}{A string for the header of the column of position (base pair) information in data.}

\item{p}{A string for the header for the column of P-value in data.}

\item{main}{A string for the Title of the plot.}

\item{suggest_line}{Set FALSE to remove the suggestive threshold line.}

\item{gws_line}{Set FALSE to remove the genome-wide significant line.}

\item{color}{A string vector containg the colors to be used.}

\item{chr_build}{A string describing what Genetic Reference Consortium build is used. It has to be either "GRCh37" or "CRCh38".}

\item{yscale}{A numeric value for y scale.}

\item{xlab_all}{Set TRUE to show labels for all chromosomes.}

\item{turbo}{Set TRUE to remove all rows with P <= 0.1.}
}
\description{
Creates Manhattan plots from GWAS summaries.
}

